Add function to call 'checkcode' in MATLAB files
The native support of matlab-mode for mlint does not work in newer
versions of MATLAB because they do not provide the mlint command line
program anymore.

Therefore we create here the 'darlan/matlab-checkcode-current-file' that
uses the capabilities provided by matlab-mode to run the "checkcode"
MATLAB command in the file associated with the current buffer.